[1.0.33] CONDITION: developing  - Fixed bur with the quantize in UniswapV2Swap.py  - some minor change

+++ src/DataFetcher/api_clients/Raydium/RaydiumMetadataTypingDict.py
@@ -1,4 +1,4 @@
 from src.DataFetcher.api_clients.BaseMetadataTypingDict import BaseMetadata
 
 class RaydiumMetadata(BaseMetadata):
-    feeRate: int
+    feeRate: float
